refactory rpc_service, split functions into two parts for http usage
