+++ client/json-rpc/rpc_service.c
@@ -27,7 +27,7 @@
 
 #include "../utils/log.h"
 #include "../system.h"
-#include "rpc_procedures.h"
+#include "rpc_procedure.h"
 
 /*
  *
@@ -50,9 +50,6 @@
 
 const uint32_t RPC_SERVER_PORT = 7677; //default http json-rpc port 7677
 
-static struct xdag_rpc_procedure *g_procedures;
-static int g_procedure_count = 0;
-
 static int send_response(struct xdag_rpc_connection * conn, char *response) {
 	int fd = conn->fd;
 	xdag_debug("JSON Response:\n%s\n", response);
@@ -60,146 +57,6 @@ static int send_response(struct xdag_rpc_connection * conn, char *response) {
 	write(fd, "\n", 1);
 	return 0;
 }
-
-static int send_error(struct xdag_rpc_connection * conn, int code, char* message, cJSON * id, char *version) {
-	int return_value = 0;
-	cJSON *result_root = cJSON_CreateObject();
-	cJSON *error_root = cJSON_CreateObject();
-	cJSON_AddNumberToObject(error_root, "code", code);
-	cJSON_AddStringToObject(error_root, "message", message);
-	cJSON_AddItemToObject(result_root, "error", error_root);
-	cJSON_AddItemToObject(result_root, "id", id);
-	
-	if(strcmp(version, "2.0")==0) {
-		cJSON_AddItemToObject(result_root, "jsonrpc", cJSON_CreateString(version));
-	} else if(strcmp(version, "1.1")==0) {
-		cJSON_AddItemToObject(result_root, "version", cJSON_CreateString(version));
-	}
-	
-	char * str_result = cJSON_Print(result_root);
-	return_value = send_response(conn, str_result);
-	free(str_result);
-	cJSON_Delete(result_root);
-	free(message);
-	return return_value;
-}
-
-static int send_result(struct xdag_rpc_connection * conn, cJSON * result, cJSON * id, char *version) {
-	int return_value = 0;
-	cJSON *result_root = cJSON_CreateObject();
-	if(result) {
-		cJSON_AddItemToObject(result_root, "result", result);
-	}
-	cJSON_AddItemToObject(result_root, "error", NULL);
-	cJSON_AddItemToObject(result_root, "id", id);
-	
-	if(strcmp(version, "2.0")==0) {
-		cJSON_AddItemToObject(result_root, "jsonrpc", cJSON_CreateString(version));
-	} else if(strcmp(version, "1.1")==0) {
-		cJSON_AddItemToObject(result_root, "version", cJSON_CreateString(version));
-	}
-	
-	char * str_result = cJSON_Print(result_root);
-	return_value = send_response(conn, str_result);
-	free(str_result);
-	cJSON_Delete(result_root);
-	return return_value;
-}
-
-static void xdag_rpc_service_procedure_destroy(struct xdag_rpc_procedure *procedure)
-{
-	if(procedure->name){
-		free(procedure->name);
-		procedure->name = NULL;
-	}
-	if(procedure->data){
-		free(procedure->data);
-		procedure->data = NULL;
-	}
-}
-
-int xdag_rpc_service_register_procedure(xdag_rpc_function function_pointer, char *name, void * data) {
-	int i = g_procedure_count++;
-	if(!g_procedures) {
-		g_procedures = malloc(sizeof(struct xdag_rpc_procedure));
-	} else {
-		struct xdag_rpc_procedure * ptr = realloc(g_procedures, sizeof(struct xdag_rpc_procedure) * g_procedure_count);
-		if(!ptr) {
-			xdag_err("rpc server : realloc failed!");
-			return -1;
-		}
-		g_procedures = ptr;
-	}
-	
-	if((g_procedures[i].name = strdup(name)) == NULL) {
-		return -1;
-	}
-	
-	g_procedures[i].function = function_pointer;
-	g_procedures[i].data = data;
-	return 0;
-}
-
-int xdag_rpc_service_unregister_procedure(char *name) {
-	int i, found = 0;
-	if(g_procedures){
-		for (i = 0; i < g_procedure_count; i++){
-			if(found) {
-				g_procedures[i-1] = g_procedures[i];
-			} else if(!strcmp(name, g_procedures[i].name)){
-				found = 1;
-				xdag_rpc_service_procedure_destroy(&(g_procedures[i]));
-			}
-		}
-		if(found){
-			g_procedure_count--;
-			if(g_procedure_count){
-				struct xdag_rpc_procedure * ptr = realloc(g_procedures, sizeof(struct xdag_rpc_procedure) * g_procedure_count);
-				if(!ptr){
-					xdag_err("rpc server : realloc failed!");
-					return -1;
-				}
-				g_procedures = ptr;
-			}else{
-				g_procedures = NULL;
-			}
-		} else {
-			xdag_err("rpc server : procedure '%s' not found\n", name);
-		}
-	} else {
-		xdag_err("rpc server : procedure '%s' not found\n", name);
-		return -1;
-	}
-	return 0;
-}
-
-static int invoke_procedure(struct xdag_rpc_connection * conn, char *name, cJSON *params, cJSON *id, char *version) {
-	cJSON *returned = NULL;
-	int procedure_found = 0;
-	struct xdag_rpc_context ctx;
-	ctx.error_code = 0;
-	ctx.error_message = NULL;
-	int i = g_procedure_count;
-	while (i--) {
-		if(!strcmp(g_procedures[i].name, name)) {
-			procedure_found = 1;
-			ctx.data = g_procedures[i].data;
-			returned = g_procedures[i].function(&ctx, params, id, version);
-			break;
-		}
-	}
-	
-	if(!procedure_found) {
-		return send_error(conn, RPC_METHOD_NOT_FOUND, strdup("Method not found."), id, version);
-	} else {
-		if(ctx.error_code) {
-			return send_error(conn, ctx.error_code, ctx.error_message, id, version);
-		} else {
-			return send_result(conn, returned, id, version);
-		}
-	}
-}
-
 /* create xdag connection */
 static struct xdag_rpc_connection* create_connection(int fd, const char* req_buffer, size_t len)
 {
@@ -228,68 +85,22 @@ static void close_connection(struct xdag_rpc_connection* conn)
 	free(conn);
 }
 
-/* handle connection */
-static void *rpc_handle_connection(void *arg)
+/* handle rpc request thread */
+static void* rpc_handle_thread(void *arg)
 {
 	struct xdag_rpc_connection* conn = (struct xdag_rpc_connection *)arg;
-	int ret = 0;
-	cJSON *root;
-	const char *end_ptr = NULL;
 	
-	if((root = cJSON_ParseWithOpts(conn->buffer, &end_ptr, 0)) != NULL) {
-		
-		char * str_result = cJSON_Print(root);
-		xdag_debug("Valid JSON Received:\n%s\n", str_result);
-		free(str_result);
-		
-		if(root->type == cJSON_Object) {
-			cJSON *method, *params, *id, *verjson;
-			char version[8] = "1.0";
-			method = cJSON_GetObjectItem(root, "method");
-			
-			verjson = cJSON_GetObjectItem(root, "jsonrpc"); /* rpc 2.0 */
-			if(!verjson) {
-				verjson = cJSON_GetObjectItem(root, "version"); /* rpc 1.1 */
-			}
-			
-			if(verjson) {
-				strcpy(version, verjson->valuestring);
-			}
-			
-			if(method != NULL && method->type == cJSON_String) {
-				params = cJSON_GetObjectItem(root, "params");
-				if(params == NULL|| params->type == cJSON_Array || params->type == cJSON_Object) {
-					id = cJSON_GetObjectItem(root, "id");
-					if(id == NULL|| id->type == cJSON_String || id->type == cJSON_Number) {
-						//We have to copy ID because using it on the reply and deleting the response Object will also delete ID
-						cJSON * id_copy = NULL;
-						if(id != NULL) {
-							id_copy = (id->type == cJSON_String) ? cJSON_CreateString(id->valuestring):cJSON_CreateNumber(id->valueint);
-						}
-						xdag_debug("Method Invoked: %s\n", method->valuestring);
-						if(invoke_procedure(conn, method->valuestring, params, id_copy, version)) {
-							xdag_err("rpc service: invoke method %s failed", method->valuestring);
-						}
-						cJSON_Delete(root);
-						close_connection(conn);
-						return 0;
-					}
-				}
-			}
-		}
-		cJSON_Delete(root);
-		if(send_error(conn, RPC_PARSE_ERROR, strdup("Request parse error."), 0, "2.0")) { //use rpc 2.0 as default version
-			xdag_err("rpc service: send error failed.");
-		}
-		close_connection(conn);
-		return 0;
-	} else {
-		if(send_error(conn, RPC_PARSE_ERROR, strdup("Request parse error."), 0, "2.0")) { //use rpc 2.0 as default version
-			xdag_err("rpc service: send error failed.");
-		}
-		close_connection(conn);
-		return 0;
+	cJSON * result = xdag_rpc_handle_request(conn->buffer);
+	char *response = cJSON_Print(result);
+	send_response(conn, response);
+	free(response);
+	
+	if(result) {
+		cJSON_Delete(result);
 	}
+	close_connection(conn);
+	
+	return 0;
 }
 
 #define BUFFER_SIZE 2048
@@ -342,7 +153,7 @@ static void *rpc_service_thread(void *arg)
 		struct xdag_rpc_connection * conn = create_connection(client_fd, req_buffer, len);
 		
 		pthread_t th;
-		int err = pthread_create(&th, 0, rpc_handle_connection, conn);
+		int err = pthread_create(&th, 0, rpc_handle_thread, conn);
 		if(err) {
 			xdag_err("rpc service : create thread failed. error : %s", strerror(err));
 			close_connection(conn);
